backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy>=2.0.0
alembic>=1.14.0
psycopg2-binary>=2.9.9
celery[redis]>=5.4.0
pydantic>=2.6.0
pydantic-settings>=2.5.0
httpx>=0.27.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
google-genai>=1.0.0
openai>=1.30.0
anthropic>=0.30.0
jinja2>=3.1.3
cryptography>=44.0.0
slowapi>=0.1.9
croniter>=1.4.0
langfuse>=4.0.0
mcp[cli]>=1.20.0
authlib>=1.3.0
redis>=5.0.0
